Ben wartości ruchu 1.0

+++ Icy Tower/Icy Tower/Map.h
@@ -29,7 +29,7 @@ public:
 	int rightBounceCounter;
 	int BounceTimer;
 	int speedBoost;
-	int verticalSpeedBoost;
+	float verticalSpeedBoost;
 	int bitmapMove;
 
 	Map();
